+++ app/Http/Controllers/BookTypeController.php
@@ -17,6 +17,8 @@ final class BookTypeController extends Controller
 {
     public function index(BookTypeServiceInterface $bookTypeService): JsonResponse
     {
+        Gate::authorize('viewAny', BookType::class);
+
         $bookTypes = $bookTypeService->getAll();
         $data = ['bookTypes' => BookTypeResource::collection($bookTypes)];
 
@@ -42,6 +44,8 @@ final class BookTypeController extends Controller
 
     public function show(BookType $bookType): JsonResponse
     {
+        Gate::authorize('view', $bookType);
+
         $data = ['bookType' => new BookTypeResource($bookType)];
 
         return (new SuccessResource($data))->response()->setStatusCode(Response::HTTP_OK);
@@ -64,8 +68,14 @@ final class BookTypeController extends Controller
         return (new SuccessResource($data))->response()->setStatusCode(Response::HTTP_OK);
     }
 
-    public function destroy(BookTypeServiceInterface $bookTypeService, BookType $bookType): JsonResponse
+    public function destroy(BookTypeServiceInterface $bookTypeService, string $bookTypeId): JsonResponse
     {
+        $bookType = $bookTypeService->get($bookTypeId);
+
+        if (!$bookType) {
+            return (new SuccessResource())->response()->setStatusCode(Response::HTTP_ACCEPTED);
+        }
+
         Gate::authorize('delete', $bookType);
 
         if (!$bookTypeService->delete($bookType)) {
